requirement.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.8.6
django-haystack==2.4.0
django-taggit==0.17.1
idna==2.6
Markdown==2.6.2
pysolr==3.3.2
pytz==2017.3
requests==2.18.4
urllib3==1.22
